+++ src/Urls/Models/UrlRecord.php
@@ -50,7 +50,7 @@ class UrlRecord extends Model
 
     public function model(): \Illuminate\Database\Eloquent\Relations\MorphTo
     {
-        return $this->morphTo('model');
+        return $this->morphTo('model')->withoutGlobalScopes();
     }
 
     /**
